Updated SAGE to properly only save the grids.  Also updated the relevant read scripts.  Finally updated the Mvir-Ngamma plotting function to save the mean/standard deviations to file.  Have started a Python script to bin the Simfast Halos based on the fits (much easier than stupid C).

+++ core_allvars.h
@@ -209,8 +209,8 @@ struct GALAXY
   float *GridSFR;
   float *GridZ;
   float *GridCentralGalaxyMass;
-  double *MfiltGnedin;
-  double *MfiltSobacchi;
+  float *MfiltGnedin;
+  float *MfiltSobacchi;
   float *EjectedFraction;
   int *LenHistory;
   double *Stars;
